fix: move keyboard code to one file, slow down page 2, fix lint issues
