+++ src/keyboard.js
@@ -1,0 +1,13 @@
+window.addEventListener("keydown", (event) => {
+	if (event.key === "ArrowRight") {
+		const form = document.querySelector("form");
+		if (form) {
+			const action = form.getAttribute("action");
+			if (action) {
+				window.location.href = action;
+			}
+		}
+	} else if (event.key === "ArrowLeft") {
+		window.history.back();
+	}
+});
