+++ src/page-3/page-3.js
@@ -1,16 +1,4 @@
-window.addEventListener("keydown", (event) => {
-	if (event.key === "ArrowRight") {
-		const form = document.querySelector("form");
-		if (form) {
-			const action = form.getAttribute("action");
-			if (action) {
-				window.location.href = action;
-			}
-		}
-	} else if (event.key === "ArrowLeft") {
-		window.history.back();
-	}
-});
+import "../keyboard.js";
 
 const names = [
 	{ name: "Joakim", type: "gjensidige", country: "Sweden", team: undefined },
