+++ src/page-7/page-7.js
@@ -1,26 +1,15 @@
-window.addEventListener("keydown", (event) => {
-	if (event.key === "ArrowRight") {
-		const form = document.querySelector("form");
-		if (form) {
-			const action = form.getAttribute("action");
-			if (action) {
-				window.location.href = action;
-			}
-		}
-	} else if (event.key === "ArrowLeft") {
-		window.history.back();
-	}
-});
+import "../keyboard.js";
 
 window.onload = () => {
 	const digits = ["1", "1", "1"];
 	const classes = ["left", "middle", "right"];
 
-	digits.forEach((digit, index) => {
+	for (let i = 0; i < digits.length; i++) {
+		const digit = digits[i];
 		const element = document.createElement("div");
-		element.className = "digit " + classes[index];
+		element.className = `digit ${classes[i]}`;
 		element.textContent = digit;
-		element.style.left = `calc(50% + ${(index - 1) * 8}rem)`;
+		element.style.left = `calc(50% + ${(i - 1) * 8}rem)`;
 		document.body.appendChild(element);
-	});
+	}
 };
